Update patch set 5
Patch Set 5:

My two cents, unrelated to this specific patch, but regarding the whole crypto/aes folder structure.
When I added the amd64 implementation of gcm, it added one .s file and one .go file. The .go file was necessary at the time. But since then every platform specific optimization that was merged added two files each, for no apparent reason. The .go files are almost identical, and what little differences there is between them, could be resolved in the assembly file, instead adding more junk to the folder.
I think, in general, the various .go files should be merged into one consistent interface to the assembly, and the assembly modified accordingly.
Again, this is not to the authors of this specific patch, but suggestion to the maintainers. It starts to look a lot like openssl.

Patch-set: 5
CC: Gerrit User 7601 <7601@62eb7196-b449-3ce5-99f1-c037f21e1705>