Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > I don't understand why this was rolled back. Where is the context? How is it that this went through with no apparent input from Robert and no linked GitHub issue?
> 
> The GitHub issue is golang.org/issue/18264. I had thought that clicking "Revert" would put some bookkeeping to make it easier to link back to the original CL.
> 
> For "no apparent input from Robert"; that is an oversight by me.

FWIW, I did see the CL fly by and I did see Brad's +2. I was ok with the original CL (removing the leading newlines after opening braces) but it does lead to situations where we now have trailing empty lines before a closing brace (and sometimes the empty lines were places for symmetry on purpose). So fixing the one thing doesn't necessarily improve the layout in general. I was ok with the revert based on that observation (also made by Joe).

In summary, I don't feel strongly about this, but it does seem that in the spirit of reducing unnecessary churn (in released gofmt formatting) it may be better to wait with this for the more complete fix which removes superfluous newlines after { and before }.

Patch-set: 2
Reviewer: Gerrit User 5210 <5210@62eb7196-b449-3ce5-99f1-c037f21e1705>