Update patch set 2
Patch Set 2:

Thanks for the context. I've removed this change from the release notes. If the point of the rollback is to wait to a later release to ensure that blank lines at the start and end of the function are handled the same way, that sounds fine. 

Note that there are two other changes to gofmt in this release, so I'm not sure the overall pain is reduced by rolling this back, but I understand wanting to wait and have a consistent change.

Patch-set: 2