Update patch set 2
Patch Set 2:

> Each of the individual tests already had golden values to lock in
 > the encoded output. You can just delete those if you want to go
 > with your version.

I kept them for now. I wouldn't be upset if someone else deleted the golden tests in the other directories, but I'm not going to bother to do that myself.

For what it's worth, the crypto/sha1 test tables made clear that information was leaking across h.Reset, which would have been a security problem. That's fixed. Also the crypto/sha256 and crypto/sha512 tables apparently continue to work even with the modification to the encoding in this CL, suggesting that those tests are not terribly precise (and they had the same security problem, it just wasn't evident in the golden data).

Patch-set: 2
Reviewer: Gerrit User 5056 <5056@62eb7196-b449-3ce5-99f1-c037f21e1705>