cmd/compile: fix bug with DWARF abstract functions
Change DWARF generation to use a different name when emitting an
abstract subprogram DIE for a function within its own package, as
opposed to an abstract subprogram DIE corresponding to an imported
inlined routine. This is to accommodate the fact that in the first
case the compiler needs to emit DWARF entries for synthesized return
temps (~r0, ~r1, etc), whereas in the second case we don't want DWARF
for the return temps.

Fixes #23046.

Change-Id: I5876bde6ea73ec2fe0651f084f52bea674610112