Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> 
> This is not a new optimization. It’s  switching back to the same exact code generation that 1.9 did, after it got broken (as a side effect) during the 1.10 cycle. If the generates code worked for 1.9, I can’t see how it can break now; on the contrary, by not merging this, algorithms that rely on bits.Ones() can get slower for no reason.

Well, yes, the generated code goes back to something we used to generate.  But that's not the major worry; the danger is that this CL doesn't actually do that faithfully.  Not that I'm particularly worried about this CL, as it is pretty simple.  But the principle is that at this stage of the release cycle, we only fix critical bugs.  This is not a correctness bug, and the performance bug is very, very minor.

Patch-set: 1
Reviewer: Gerrit User 5200 <5200@62eb7196-b449-3ce5-99f1-c037f21e1705>