go/types: delay type-checking of closures
R=go1.11

Functions (at the package level) were collected and their bodies
type-checked after all other package-level objects were checked.
But closures where type-checked right when they were encountered
so that they could see the correct, partially populated surrounding
scope, and also to mark variables of the surrounding function as
used.

This apprach, while simple, breaks down in esoteric cases where
a closure appears inside the declaration of an object that its
body depends on: If the body is type-checked before the object
is completely set up, the closure may use incomplete data
structures, possibly leading to spurious errors.

This change postpones type-checking of closures to later; after
the current expression or statement, but before any changes to
the enclosing scope (so that the delayed type-checking sees the
correct scope contents).

The new mechanism is general and now is also used for other
(non-function) delayed checks.

Fixes #22992.

Change-Id: Ic95f709560858b4bdf8c645be70abe4449f6184d