go/internal/srcimporter: fix misuse of build.Import and build.ImportDir
Current implementation is doing following jobs in importing the package.

1. import the package by build.Import(Dir) with build.FindOnly.
2. re-import the package as a local-import by build.ImportDir with the
pkgdir which is got from step 1.

Step 2 is unnecessary unless we specify build.FindOnly on step 1.

Now, it turns out that treating every import as a local-import is not
only redundant but also causes several problems.

So, fix it.

Fixes #23092
Fixes #24392

Change-Id: I9e900538b365398751bace56964495c5440ac4ae