Update patch set 9
Patch Set 9:

> Patch Set 9:
> 
> > @iant, thank you for the suggestion on the circular buffer. I got time this evening to implement it and the latest patchset has it. However, I've encountered a writebarrier error:
> > ./traceback.go:415:14: write barrier prohibited by caller; gentraceback
> > 	/Users/emmanuelodeke/go/src/go.googlesource.com/go/src/runtime/proc.go:3699:19: called by sigprof
> > when saving the frame while printing. If you could please help me take a look at this, that'd be great. However, patchset 8 is rudimentary but does the work inefficiently, plus also it has the suggestions for the tests.
> 
> gentraceback is called by the signal handler, so it must not use any write barriers.  It also can't allocate any memory.  I'm not sure whether the compiler is smart enough to see that the &frameSave doesn't escape.  Also, this means that your frameSave type must always live on the stack or must not contain any pointers.  It may work to mark it notinheap, I'm not sure.
> 
> I guess this is going to be more complicated than I thought.  The call from the signal handler doesn't print anything, but the compiler doesn't know that.  And it's going to be hard to pull apart gentraceback without duplicating code that we don't want to duplicate.  I'm not sure of the best way to move forward.

Overall though the idea is great, its just that it turns hairy but I believe that with some more experimentation we'll be able to get it optimal :) In regards to a way forward, without speaking out of turn, I'd kindly suggest that we start with the basic/non-optimal solution, give users the benefits of having a nicer stack trace, mark a TODO and we can perhaps collaborate on it in the next Go cycles, and we already have a basis CL patchset right here for reference in the future. What do you think?

Patch-set: 9
Reviewer: Gerrit User 5137 <5137@62eb7196-b449-3ce5-99f1-c037f21e1705>