Add capabilities to inheritable set before adding to ambient set.
The code previously simply added the capabilities to the ambient set.  This
works if you started life as a process with all capabilities in all sets
(i.e. if you were owned by root) but if you're creating a user namespace in
which you're not root inside that namespace, you have all capabilities, but
your inheritable set is empty and the ambient code will fail.  If you need
the child process you exec to still have some capabilities, there is no way
to make that happen at all without this fix.

Change-Id: I38ed145759b9fd115b8c58a9c7a683e3925da281