archive/tar: allow legacy approach to global PAX headers
DO NOT SUBMIT

Prior to Go 1.10, there was no support for global PAX headers,
so users who desired esoteric features were forced to
manually encode a raw PAX header themselves using the 'g' typeflag.
This approach is naturally brittle when proper support is added
to the tar package.

In order to ensure that the manual approach continues to work,
we special-case the condition where Header.Size > 0, which indicates
the intention to write the header manually, but do assert that
any PAX related features (PAXRecords, Xattrs, and paxHdrs) are not used
(otherwise it is non-sensible what happens when mixing local and global
header records).

Change-Id: I5d23a795d83a35cca1ba9012e614fa491235dd3f