Update patch set 9
Patch Set 9:

> Patch Set 9:
> 
> > Patch Set 8:
> > 
> > > Patch Set 7:
> > > 
> > > Where is check.pos assigned? The only place I can see is when the type checker is invoked through Eval, which is a very esoteric case. pos will be NoPos most of the time in practice, except when you Eval an expression.  If that expression contains a function literal, various checks that should be independent of the client-specified pos value (e.g. "not in scope at return", Var.used) will now depend on it.
> > 
> > It's only set by Eval. I reverted the chances where the meaning changed.
> > 
> > PTAL.
> 
> I suspect there are preexisting problems with Eval that remain, but I haven't tested.  For example, every Ident is resolved relative to check.pos even if it appears in a narrower scope with a different binding. For example, the x in Eval("func(x int) { _ = x }", pos) will appear to resolve not to the parameter but within the environment at pos.
> 
> Also, how does LookupParent interact with your other CL that delays function body processing? Consider:
> 
> 	const (
> 		a = unsafe.Sizeof(func() { _ = b })
> 		b = 1
> 	)
> 
> At package level this out-of-order const declaration is legal, but within a function body the forward reference to b is forbidden.  Does the deferred processing of the function body cause it to be resolved in an environment that defines b?  (I suspect that the code used to rely on processing happening essentially left-to-right, which is no longer the case due to deferral.)

There may be issues to Eval that are pre-existing; this CL is not trying to address those, and at this point this CL should not change go/types behavior, hopefully.

Regarding the delayed function literal processing: My original intent was to use the literal's position to control its scope access even when evaluated in a delayed fashion. It did work but I decided to not rely on positions for it after all because one of the principles I've tried to stick by for go/types is that it should work correctly even if we don't have any (or incorrect) position information (e.g. we use an explicit object numbering to allow sorting in source order w/o the need of positions).

The delayed processing is delayed at most to the end of a statement (inside a function) to make sure that we don't see new objects in scope that shouldn't be. There's special code for DeclStmt nodes to make sure that we resolve delayed function literals appearing in initialization expressions of const/var decls _before_ the respective const/var names are in scope. For local type declarations (e.g., type A [unsafe.sizeof(func(){...})]byte) it's ok
process the function literals before the next statement since type names are immediately visible.

Package-level function literals are treated like ordinary functions, they are processed at the end.

Patch-set: 9
Reviewer: Gerrit User 5210 <5210@62eb7196-b449-3ce5-99f1-c037f21e1705>