Update patch set 9
Patch Set 9: Code-Review+2

> Patch Set 9:
>  
> There may be issues to Eval that are pre-existing; this CL is not trying to address those, and at this point this CL should not change go/types behavior, hopefully.

I agree this change should be behavior-preserving, so LGTM.  It does sweep the details of check.pos under the rug a little, which might make it easier to make mistakes in future.

 
> go/types ... should work correctly even if we don't have any (or incorrect) position information

Good.


> For local type declarations (e.g., type A [unsafe.sizeof(func(){...})]byte) it's ok
> process the function literals before the next statement since type names are immediately visible.

Do you have a test for that change that exercises the case in which a function body in the middle of a type decl "destructures" the named type that is still only a forward reference?   For example, this func wants to access the element type of channel type A:

	type A chan [unsafe.Sizeof(func() { <-make(A) })]int

Alternatively: <-A(nil).

I think gc handles this by binding an incomplete type name to a special forward reference type that doesn't work with make(T), T(nil), etc.

Patch-set: 9
Reviewer: Gerrit User 5195 <5195@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=+2