go/doc, cmd/doc: record/display some version info
Currently, it is difficult to tell when a function, method, or type was
added to the standard library.

This change introduces mkapi.go, which reads api/go*.txt files to gather
information about which version of Go introduces any particular
function, method, or type and serializes this data into a nested map
structure with the following hierarchy:

package -> (type|func|method) -> name -> introduced version

In the case of methods, both the receiver and the function name are
encoded in the "name" portion of this hierarchy, separated by a period
character.

This change does not address constants or variables, and completely
ignores the syscall package. The former are future work, the latter is
likely an exercise in futility. In all cases, this is because the story
around displaying the version information is not well developed.

Separately, cmd/doc is modified to look up and output this version
information when it is available. For any package outside of the
standard library, the output of go doc does not change at all. Summary
output appends "(since <version>)", documentation on individual
functions / methods / types adds a "Since <version>" line to the top of
the output.

Fixes #5778

Change-Id: Ib4f7026a98d1f9c9e8e1155551ae9be179a8e8b7