cmd/compile/internal/syntax: remove dependency on cmd/internal/src
DO NOT COMMIT. DO NOT REVIEW.

Experiment in progress.

For dependency reasons, the data structure implementing source
positions in the compiler is in cmd/internal/src. It contains
highly compiler specific details (e.g. inlining index).

This change introduces a parallel but simplified position
representation, defined in the syntax package, which removes
that package's dependency on cmd/internal/src, and also removes
the need to deal with certain filename-specific operations
(defined by the needs of the compiler) in the syntax package.
As a result, the syntax package becomes again a compiler-
independent, stand-alone package that at some point might
replace (or augment) the existing top-level go/* syntax-related
package.

This change comes at the cost of another position translation
step, but that step is cheap because it only needs to do real
work if the position base changed (i.e., if there is a new file,
or new line directive).

Change-Id: I19edb53dd9675af57f7122cb7dba2a6d8bdcc3da