Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> If the expression is long and complex, the syntax error can still be a bit confusing. But I do agree that pointing out the assignment does give a hint, which is better than what we had before.
> 
> By the way, no test?

It should have a test, ideally as part of the syntax package, but writing a little versatile test framework for that package is an open issue. I "tested" manually for now. We will have an extensive test suite eventually.

(One could add a test to test/fixedbugs but those have to also accept gccgo errors and are sometimes not as precise (position) as desired. This is such a trivial change I didn't bother.)

Patch-set: 1
Reviewer: Gerrit User 5210 <5210@62eb7196-b449-3ce5-99f1-c037f21e1705>